chat SDK version 2

+++ ChatSdk/src/main/java/com/olashiku/chatsdk/model/request/new_message/NewMessageRequest.kt
@@ -7,5 +7,5 @@ data class NewMessageRequest(
     val msgPlatform: String,
     val msgReceiver: MsgReceiver? = null,
     val msgSender: MsgSender,
-    val msgTimeStamp: Int
+    val msgTimeStamp: Long
 )
